refactor: update image handling and redirect configuration; remove unused components
